requirements.txt
matplotlib==3.5.1
networkx==2.7.1
numpy==1.21.5
palettable==3.3.0
pandas==1.4.2
scikit_learn==1.2.2
stellargraph==1.2.1
tensorflow-gpu==2.6.0
torch==1.12.0+cu116
